fix for remove_files task
The info.json files will have paths in the old location (src/file_or_folder_to_remove) vs the new location (buildsrc/plugin_slug/file_or_folder_to_remove).  The fix that allows keeping info.json as is, is to assume that info.json either has the path defined relative to the plugin install in the buildmachine OR has `src/` prefixed and that will get replaced with the actual location.